--- BE/src/main/java/roomescape/domain/ReservationTimeRepository.java
@@ -1,11 +1,0 @@
-package roomescape.domain;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface ReservationTimeRepository {
-    ReservationTime save(ReservationTime reservation);
-    Optional<ReservationTime> findById(Long id);
-    List<ReservationTime> findAll();
-    void deleteById(Long id);
-}
