--- BE/src/main/java/roomescape/domain/Theme.java
@@ -1,67 +1,0 @@
-package roomescape.domain;
-
-public class Theme {
-
-    private final Long id;
-    private final String name;
-    private final String description;
-    private final String thumbnailUrl;
-
-    private Theme (Long id, String name, String description, String thumbnailUrl) {
-        this.id = id;
-        this.name = name;
-        this.description = description;
-        this.thumbnailUrl = thumbnailUrl;
-    }
-
-    public static Theme create(String name, String description, String thumbnailUrl) {
-        return new Theme(null, name, description, thumbnailUrl);
-    }
-
-    public static Theme createRow(Long id, String name, String description, String thumbnailUrl) {
-        return new Theme(id, name, description, thumbnailUrl);
-    }
-
-    public Theme appendId(Long id) {
-        return new Theme(id, name, description, thumbnailUrl);
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public String getThumbnailUrl() {
-        return thumbnailUrl;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (!(o instanceof Theme that)) {
-            return false;
-        }
-        if (this.id == null || that.id == null) {
-            return false;
-        }
-        return this.id.equals(that.id);
-    }
-
-    @Override
-    public int hashCode() {
-        if (id != null) {
-            return id.hashCode();
-        } else {
-            return System.identityHashCode(this);
-        }
-    }
-}
