--- BE/src/main/java/roomescape/domain/ThemeRepository.java
@@ -1,11 +1,0 @@
-package roomescape.domain;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface ThemeRepository {
-    Theme save(Theme theme);
-    Optional<Theme> findById(Long id);
-    List<Theme> findAll();
-    void deleteById(Long id);
-}
