--- BE/src/main/java/roomescape/domain/ThemeSortType.java
@@ -1,5 +1,0 @@
-package roomescape.domain;
-
-public enum ThemeSortType {
-    POPULAR
-}
