--- BE/src/main/java/roomescape/presentation/dto/ReservationResponse.java
@@ -1,21 +1,0 @@
-package roomescape.presentation.dto;
-
-import roomescape.domain.Reservation;
-
-public record ReservationResponse(
-        Long id,
-        String name,
-        String date,
-        ReservationTimeResponse time,
-        ThemeResponse theme
-) {
-    public static ReservationResponse from(Reservation reservation) {
-        return new ReservationResponse(
-                reservation.getId(),
-                reservation.getName(),
-                reservation.getDate().toString(),
-                ReservationTimeResponse.from(reservation.getTime()),
-                ThemeResponse.from(reservation.getTheme())
-        );
-    }
-}
