+++ BE/src/main/java/roomescape/reservation/domain/Reservation.java
@@ -1,0 +1,77 @@
+package roomescape.reservation.domain;
+
+import java.time.LocalDate;
+import roomescape.reservationTime.domain.ReservationTime;
+import roomescape.theme.domain.Theme;
+
+public class Reservation {
+
+    private final Long id;
+    private final String name;
+    private final LocalDate date;
+    private final ReservationTime time;
+    private final Theme theme;
+
+    private Reservation(Long id, String name, LocalDate date, ReservationTime time, Theme theme) {
+        this.id = id;
+        this.name = name;
+        this.date = date;
+        this.time = time;
+        this.theme = theme;
+    }
+
+    public static Reservation create(String name, LocalDate date, ReservationTime time, Theme theme) {
+        return new Reservation(null, name, date, time, theme);
+    }
+
+    public static Reservation createRow(Long id, String name, LocalDate date, ReservationTime time, Theme theme) {
+        return new Reservation(id, name, date, time, theme);
+    }
+
+    public Reservation appendId(Long id) {
+        return new Reservation(id, name, date, time, theme);
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public LocalDate getDate() {
+        return date;
+    }
+
+    public ReservationTime getTime() {
+        return time;
+    }
+
+    public Theme getTheme() {
+        return theme;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (!(o instanceof Reservation that)) {
+            return false;
+        }
+        if (this.id == null || that.id == null) {
+            return false;
+        }
+        return this.id.equals(that.id);
+    }
+
+    @Override
+    public int hashCode() {
+        if (id != null) {
+            return id.hashCode();
+        } else {
+            return System.identityHashCode(this);
+        }
+    }
+}
