+++ BE/src/main/java/roomescape/reservationTime/application/ReservationTimeService.java
@@ -1,0 +1,61 @@
+package roomescape.reservationTime.application;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+import java.util.List;
+import java.util.Set;
+import java.util.stream.Collectors;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import roomescape.global.exception.ReservationTimeErrorCode;
+import roomescape.global.exception.customException.BusinessException;
+import roomescape.reservation.domain.Reservation;
+import roomescape.reservation.domain.ReservationRepository;
+import roomescape.reservationTime.domain.ReservationTime;
+import roomescape.reservationTime.domain.ReservationTimeRepository;
+
+@Service
+public class ReservationTimeService {
+
+    private final ReservationTimeRepository reservationTimeRepository;
+    private final ReservationRepository reservationRepository;
+
+    public ReservationTimeService(
+            ReservationTimeRepository reservationTimeRepository, ReservationRepository reservationRepository
+    ) {
+        this.reservationTimeRepository = reservationTimeRepository;
+        this.reservationRepository = reservationRepository;
+    }
+
+    @Transactional
+    public ReservationTime saveTime(LocalTime startAt) {
+        ReservationTime reservationTime = ReservationTime.create(startAt);
+        return reservationTimeRepository.save(reservationTime);
+    }
+
+    public List<ReservationTime> getTimes() {
+        return reservationTimeRepository.findAll();
+    }
+
+    public Set<Long> getBookedTimes(LocalDate date, Long themeId) {
+        if (date == null || themeId == null) {
+            return Set.of();
+        }
+        List<ReservationTime> bookedTimes = reservationRepository.findByDateAndThemeId(date, themeId)
+                .stream()
+                .map(Reservation::getTime)
+                .toList();
+
+        return bookedTimes.stream()
+                .map(ReservationTime::getId)
+                .collect(Collectors.toSet());
+    }
+
+    @Transactional
+    public void deleteTime(Long id) {
+        if (reservationRepository.existsByReservationTimeId(id)) {
+            throw new BusinessException(ReservationTimeErrorCode.RESERVATION_TIME_IN_USE);
+        }
+        reservationTimeRepository.deleteById(id);
+    }
+}
