+++ BE/src/main/java/roomescape/reservationTime/domain/ReservationTime.java
@@ -1,0 +1,52 @@
+package roomescape.reservationTime.domain;
+
+import java.time.LocalTime;
+
+public class ReservationTime {
+
+    private final Long id;
+    private final LocalTime startAt;
+
+    private ReservationTime(Long id, LocalTime startAt) {
+        this.id = id;
+        this.startAt = startAt;
+    }
+
+    public static ReservationTime create(LocalTime startAt) {
+        return new ReservationTime(null, startAt);
+    }
+
+    public static ReservationTime createRow(Long id, LocalTime startAt) {
+        return new ReservationTime(id, startAt);
+    }
+
+    public Long getId() {
+        return id;
+    }
+    public LocalTime getStartAt() {
+        return startAt;
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (!(o instanceof ReservationTime that)) {
+            return false;
+        }
+        if (this.id == null || that.id == null) {
+            return false;
+        }
+        return this.id.equals(that.id);
+    }
+
+    @Override
+    public int hashCode() {
+        if (id != null) {
+            return id.hashCode();
+        } else {
+            return System.identityHashCode(this);
+        }
+    }
+}
