+++ BE/src/main/java/roomescape/reservationTime/domain/ReservationTimeRepository.java
@@ -1,0 +1,11 @@
+package roomescape.reservationTime.domain;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface ReservationTimeRepository {
+    ReservationTime save(ReservationTime reservation);
+    Optional<ReservationTime> findById(Long id);
+    List<ReservationTime> findAll();
+    void deleteById(Long id);
+}
