+++ BE/src/main/java/roomescape/theme/domain/PopularThemeRepository.java
@@ -1,0 +1,9 @@
+package roomescape.theme.domain;
+
+import java.time.LocalDate;
+import java.util.List;
+
+public interface PopularThemeRepository {
+
+    List<Theme> findTopNByPeriod(LocalDate startAt, LocalDate endAt, ThemeSortType sortType, Long limit);
+}
