+++ BE/src/main/java/roomescape/theme/domain/ThemeRepository.java
@@ -1,0 +1,11 @@
+package roomescape.theme.domain;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface ThemeRepository {
+    Theme save(Theme theme);
+    Optional<Theme> findById(Long id);
+    List<Theme> findAll();
+    void deleteById(Long id);
+}
