+++ BE/src/main/java/roomescape/theme/domain/ThemeSortType.java
@@ -1,0 +1,5 @@
+package roomescape.theme.domain;
+
+public enum ThemeSortType {
+    POPULAR
+}
