+++ BE/src/main/java/roomescape/theme/presentation/ThemeController.java
@@ -1,0 +1,81 @@
+package roomescape.theme.presentation;
+
+import java.net.URI;
+import java.time.LocalDate;
+import java.util.List;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
+import roomescape.global.auth.Admin;
+import roomescape.theme.application.ThemeService;
+import roomescape.theme.presentation.dto.ThemeRequest;
+import roomescape.theme.presentation.dto.ThemeResponse;
+
+@RestController
+@RequestMapping("/themes")
+public class ThemeController {
+
+    private final ThemeService service;
+
+    public ThemeController(ThemeService service) {
+        this.service = service;
+    }
+
+    @Admin
+    @PostMapping
+    public ResponseEntity<ThemeResponse> saveTheme(
+            @RequestBody ThemeRequest request
+    ) {
+        ThemeResponse result = ThemeResponse.from(service.save(
+                request.name(),
+                request.description(),
+                request.thumbnail()
+        ));
+
+        URI location = ServletUriComponentsBuilder
+                .fromCurrentRequest()
+                .path("/{id}")
+                .buildAndExpand(result.id())
+                .toUri();
+
+        return ResponseEntity.created(location).body(result);
+    }
+
+    @GetMapping
+    public ResponseEntity<List<ThemeResponse>> getThemes() {
+        List<ThemeResponse> response = (service.findAll()).stream()
+                .map(ThemeResponse::from)
+                .toList();
+
+        return ResponseEntity.ok(response);
+    }
+
+    @GetMapping(params = {"sortBy", "from", "to", "limit"})
+    public ResponseEntity<List<ThemeResponse>> getTopNByPeriod(
+            @RequestParam String sortBy,
+            @RequestParam LocalDate from,
+            @RequestParam LocalDate to,
+            @RequestParam(required = false, defaultValue = "10") Long limit
+    ) {
+        List<ThemeResponse> response = (service.findTopNByPeriod(from, to, sortBy, limit)).stream()
+                .map(ThemeResponse::from)
+                .toList();
+        return ResponseEntity.ok(response);
+    }
+
+    @Admin
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteTheme(
+            @PathVariable Long id
+    ) {
+        service.deleteById(id);
+        return ResponseEntity.noContent().build();
+    }
+}
