+++ BE/src/main/java/roomescape/theme/presentation/dto/ThemeResponse.java
@@ -1,0 +1,19 @@
+package roomescape.theme.presentation.dto;
+
+import roomescape.theme.domain.Theme;
+
+public record ThemeResponse(
+        Long id,
+        String name,
+        String description,
+        String thumbnail
+) {
+    public static ThemeResponse from(Theme theme) {
+        return new ThemeResponse(
+                theme.getId(),
+                theme.getName(),
+                theme.getDescription(),
+                theme.getThumbnailUrl()
+        );
+    }
+}
